feat: Update `Expandable` to use native `<details>` element
This has some important benefits:

1. It should be covered by search etc. as this is statically rendered now and just "hidden" via CSS.
2. The browser automatically searches in it when using CMD+F - it will auto-open respective details that contain the searched content.
