+++ docs/contributing/pages/components.mdx
@@ -130,6 +130,12 @@ Render an expandable section to provide additional information to users on deman
   provide optional information that can help users be more successful.
 </Expandable>
 
+<Expandable title="Expandable with a code block">
+  ```js
+  const foo = 'bar';
+  ```
+</Expandable>
+
 ```markdown {tabTitle:Example}
 <Expandable title="Here's something worth noting">
   This is an expandable section in an `'info'` alert style.
