+++ src/components/codeBlock/code-blocks.module.scss
@@ -19,7 +19,7 @@
     font-size: 0.85rem;
     border: 0;
     border-radius: 4px;
-    margin: 0 0 1rem;
+    margin: 0;
   }
 
   /**
