+++ src/components/docPage/type.scss
@@ -136,6 +136,10 @@
     font-weight: 500;
   }
 
+  .code-tabs-wrapper:not(:last-child) {
+    margin-bottom: 1.5rem;
+  }
+
   .anchorlink,
   .anchorlink.before {
     display: flex;
