+++ src/components/expandable/style.module.scss
@@ -1,0 +1,32 @@
+.expandable-header {
+  display: flex;
+  cursor: pointer;
+  gap: 0.7rem;
+  margin-bottom: 0.3rem;
+}
+
+.expandable-body {
+  margin-left: 1.6rem;
+
+  & ul, & ol {
+    padding-left: 1.2rem;
+  }
+}
+
+.expandable {
+  &:open {
+    padding-bottom: 1rem;
+  }
+
+  &:not(:open) .expandable-header {
+    margin-bottom: 0;
+  }
+
+  &:open .expandable-icon-collapsed {
+    display: none;
+  }
+
+  &:not(:open) .expandable-icon-expanded {
+    display: none;
+  }
+}
